LPS-57392 Move java to addon: layout-controller-control-panel, change reference
